swf: Extract filters to separate files
For better code organization.
