+++ swf/src/types/color_matrix_filter.rs
@@ -1,0 +1,6 @@
+use crate::Fixed16;
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct ColorMatrixFilter {
+    pub matrix: [Fixed16; 20],
+}
