+++ swf/src/types/gradient_bevel_filter.rs
@@ -1,0 +1,15 @@
+use crate::{Fixed16, Fixed8, GradientRecord};
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct GradientBevelFilter {
+    pub colors: Vec<GradientRecord>,
+    pub blur_x: Fixed16,
+    pub blur_y: Fixed16,
+    pub angle: Fixed16,
+    pub distance: Fixed16,
+    pub strength: Fixed8,
+    pub is_inner: bool,
+    pub is_knockout: bool,
+    pub is_on_top: bool,
+    pub num_passes: u8,
+}
